+++ src/c++/perf_analyzer/load_worker.h
@@ -30,125 +30,11 @@
 #include <mutex>
 
 #include "data_loader.h"
+#include "infer_manager.h"
 #include "iworker.h"
 #include "model_parser.h"
 
 namespace triton { namespace perfanalyzer {
-
-// Tracks the data step ID for non-sequences
-class DataStepIdTracker {
- public:
-  DataStepIdTracker(size_t data_step_id) : data_step_id_(data_step_id) {}
-
-  size_t GetDataStepId() { return data_step_id_; }
-  void SetDataStepId(size_t data_step_id) { data_step_id_ = data_step_id; }
-
- private:
-  size_t data_step_id_;
-};
-
-// Holds information about the shared memory locations
-struct SharedMemoryData {
-  SharedMemoryData(
-      size_t byte_size,
-      std::unique_ptr<uint8_t, std::function<void(uint8_t*)>> data)
-      : byte_size_(byte_size), data_(std::move(data))
-  {
-  }
-
-  SharedMemoryData() {}
-
-  // Byte size
-  size_t byte_size_;
-
-  // Unique pointer holding the shared memory data
-  std::unique_ptr<uint8_t, std::function<void(uint8_t*)>> data_;
-};
-
-// Holds the status of the inflight sequence
-struct SequenceStat {
-  SequenceStat(uint64_t seq_id)
-      : seq_id_(seq_id), data_stream_id_(0), remaining_queries_(0)
-  {
-  }
-  // If paused, no more requests should be sent other than a single request to
-  // finish the active sequence
-  bool paused_ = false;
-  // The unique correlation id allocated to the sequence
-  uint64_t seq_id_;
-  // The data stream id providing data for the sequence
-  uint64_t data_stream_id_;
-  // The number of queries remaining to complete the sequence
-  size_t remaining_queries_;
-  // A lock to protect sequence data
-  std::mutex mtx_;
-};
-
-// Holds the running status of the thread.
-struct ThreadStat {
-  ThreadStat() {}
-
-  // The status of the worker thread
-  cb::Error status_;
-  // The status of the callback thread for async requests
-  cb::Error cb_status_;
-  // The statistics of the InferContext
-  std::vector<cb::InferStat> contexts_stat_;
-  // A vector of request timestamps <start_time, end_time>
-  // Request latency will be end_time - start_time
-  TimestampVector request_timestamps_;
-  // A lock to protect thread data
-  std::mutex mu_;
-};
-
-/// Wraps the information required to send an inference to the
-/// server
-struct InferContext {
-  explicit InferContext() {}
-  InferContext(InferContext&&) = delete;
-  InferContext(const InferContext&) = delete;
-  ~InferContext()
-  {
-    for (const auto input : inputs_) {
-      delete input;
-    }
-    for (const auto output : outputs_) {
-      delete output;
-    }
-  }
-  // The backend to communicate with the server
-  std::unique_ptr<cb::ClientBackend> infer_backend_;
-  // The vector of pointers to InferInput objects for all possible inputs,
-  // potentially including optional inputs with no provided data.
-  std::vector<cb::InferInput*> inputs_;
-  // The vector of pointers to InferInput objects to be
-  // used for inference request.
-  std::vector<cb::InferInput*> valid_inputs_;
-  // The vector of pointers to InferRequestedOutput objects
-  // to be used with the inference request.
-  std::vector<const cb::InferRequestedOutput*> outputs_;
-  // If not empty, the expected output data in the same order as 'outputs_'
-  std::vector<std::vector<std::pair<const uint8_t*, size_t>>> expected_outputs_;
-  // The InferOptions object holding the details of the
-  // inference.
-  std::unique_ptr<cb::InferOptions> options_;
-};
-
-/// The properties of an asynchronous request required in
-/// the callback to effectively interpret the response.
-struct AsyncRequestProperties {
-  AsyncRequestProperties() : sequence_end_(false), delayed_(true) {}
-  // The id of in the inference context which was used to
-  // send this request.
-  uint32_t ctx_id_;
-  // The timestamp of when the request was started.
-  std::chrono::time_point<std::chrono::system_clock> start_time_;
-  // Whether or not the request is at the end of a sequence.
-  bool sequence_end_;
-  // Whether or not the request is delayed as per schedule.
-  bool delayed_;
-};
-
 
 /// Abstract base class for worker threads
 ///
@@ -182,122 +68,22 @@ class LoadWorker : public IWorker {
         distribution_(distribution), wake_signal_(wake_signal),
         wake_mutex_(wake_mutex), execute_(execute)
   {
-    data_step_id_tracker_ = std::make_unique<DataStepIdTracker>(id);
+    infer_manager_ = std::make_unique<InferManager>(
+        id, async, streaming, on_sequence_model, using_json_data, batch_size,
+        backend_kind, shared_memory_type, start_sequence_id, sequence_id_range,
+        sequence_length, thread_stat, sequence_stat, data_loader, parser,
+        factory);
   }
 
   virtual ~LoadWorker() = default;
 
-  /// Creates inference input object
-  /// \param infer_input Output parameter storing newly created inference input
-  /// \param kind Backend kind
-  /// \param name Name of inference input
-  /// \param dims Shape of inference input
-  /// \param datatype Data type of inference input
-  /// \return cb::Error object indicating success or failure.
-  virtual cb::Error CreateInferInput(
-      cb::InferInput** infer_input, const cb::BackendKind kind,
-      const std::string& name, const std::vector<int64_t>& dims,
-      const std::string& datatype);
-
-  /// Helper function to prepare the InferContext for sending
-  /// inference request.
-  /// \param ctx The target InferContext object.
-  /// \return cb::Error object indicating success or failure.
-  cb::Error PrepareInfer(InferContext* ctx);
-
-  /// Helper function to prepare the InferContext for sending
-  /// inference request in shared memory. \param ctx The target
-  /// InferContext object. \return cb::Error object indicating
-  /// success or failure.
-  cb::Error PrepareSharedMemoryInfer(InferContext* ctx);
-
-  /// Updates the input data to use for inference request
-  /// \param inputs The vector of pointers to InferInput objects for all
-  /// possible inputs, potentially including optional inputs with no provided
-  /// data
-  /// \param valid_inputs The vector of pointers to InferInput objects to be
-  /// used for inference request.
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \return cb::Error object indicating success or failure.
-  cb::Error UpdateInputs(
-      const std::vector<cb::InferInput*>& inputs,
-      std::vector<cb::InferInput*>& valid_inputs, int stream_index,
-      int step_index);
-
-  /// Updates the expected output data to use for inference request. Empty
-  /// vector will be returned if there is no expected output associated to the
-  /// step.
-  /// \param outputs The vector of outputs to get the expected data
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \param data The vector of pointer and size of the expected outputs
-  /// \return cb::Error object indicating success or failure.
-  cb::Error UpdateValidationOutputs(
-      const std::vector<const cb::InferRequestedOutput*>& outputs,
-      int stream_index, int step_index,
-      std::vector<std::vector<std::pair<const uint8_t*, size_t>>>& data);
-
-  cb::Error ValidateOutputs(
-      const InferContext& ctx, const cb::InferResult* result_ptr);
-
-  void SetInferSequenceOptions(
-      const uint32_t seq_stat_index,
-      std::unique_ptr<cb::InferOptions>& options);
-  void InitNewSequence(int seq_stat_index);
-  uint64_t GetNextSeqId(int seq_stat_index);
-
-  /// Generate random sequence length based on 'offset_ratio' and
-  /// 'sequence_length_'. (1 +/- 'offset_ratio') * 'sequence_length_'
-  /// \param offset_ratio The offset ratio of the generated length
-  /// \return random sequence length
-  size_t GetRandomSequenceLength(double offset_ratio);
-
-  /// Helper function to update the inputs
-  /// \param inputs The vector of pointers to InferInput objects for all
-  /// possible inputs, potentially including optional inputs with no provided
-  /// data
-  /// \param valid_inputs The vector of pointers to InferInput objects to be
-  /// used for inference request.
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \return cb::Error object indicating success or failure.
-  cb::Error SetInputs(
-      const std::vector<cb::InferInput*>& inputs,
-      std::vector<cb::InferInput*>& valid_inputs, const int stream_index,
-      const int step_index);
-
-  /// Helper function to update the shared memory inputs
-  /// \param inputs The vector of pointers to InferInput objects
-  /// \param stream_index The data stream to use for next data
-  /// \param step_index The step index to use for next data
-  /// \return cb::Error object indicating success or failure.
-  cb::Error SetInputsSharedMemory(
-      const std::vector<cb::InferInput*>& inputs, const int stream_index,
-      const int step_index);
-
  protected:
-  // Create and initialize a new context
-  void CreateContext();
+  // FIXME
+  void PrepAndSendInferRequest(uint32_t ctx_id, bool delayed = false)
+  {
+    infer_manager_->PrepAndSendInferRequest(ctx_id, delayed);
+  }
 
-  void PrepAndSendInferRequest(uint32_t ctx_id, bool delayed = false);
-
-  /// A helper function to issue inference request to the server.
-  /// \param context InferContext to use for sending the request.
-  /// \param context_id The ID of the context
-  /// \param request_id The unique id to be associated with the request.
-  /// \param delayed Whether the request fell behind its scheduled time.
-  /// \param callback_func The callback function to use with asynchronous
-  /// request.
-  /// \param async_req_map The map from ongoing request_id to the
-  /// request information needed to correctly interpret the details.
-  /// \param thread_stat The runnning status of the worker thread
-  void SendRequest(
-      std::shared_ptr<InferContext> context, const uint32_t ctx_id,
-      const uint64_t request_id, const bool delayed,
-      cb::OnCompleteFn callback_func,
-      std::map<std::string, AsyncRequestProperties>& async_req_map,
-      std::shared_ptr<ThreadStat> thread_stat);
 
   // Detect and handle the case where this thread needs to exit
   // Returns true if an exit condition was met
@@ -305,30 +91,15 @@ class LoadWorker : public IWorker {
 
   virtual uint32_t GetSeqStatIndex(uint32_t ctx_id) = 0;
   virtual void CompleteOngoingSequences() = 0;
-  void CompleteOngoingSequence(uint32_t ctx_id, uint32_t seq_stat_index);
 
   void WaitForOngoingRequests();
-
-  // Callback function for handling asynchronous requests
-  void AsyncCallbackFuncImpl(cb::InferResult* result);
 
   // Code to execute at the end of the async callback function
   virtual void AsyncCallbackFinalize(uint32_t ctx_id) = 0;
 
-  // Function pointer to the async callback function implementation
-  std::function<void(cb::InferResult*)> async_callback_func_ = std::bind(
-      &LoadWorker::AsyncCallbackFuncImpl, this, std::placeholders::_1);
-
-  /// Update inputs based on custom json data
-  void UpdateJsonData(const uint32_t ctx_id);
-
-  /// Update inputs based on custom json data for the given sequence
-  void UpdateSeqJsonData(
-      const uint32_t ctx_id, std::shared_ptr<SequenceStat> seq_stat);
-
   virtual size_t GetNumActiveThreads() = 0;
 
-  std::unique_ptr<DataStepIdTracker> data_step_id_tracker_;
+  std::unique_ptr<InferManager> infer_manager_;
 
   // TODO REFACTOR TMA-1019 all sequence related code should be in a single
   // class. We shouldn't have to have a shared uint64 reference passed to all
@@ -381,9 +152,6 @@ class LoadWorker : public IWorker {
   const uint64_t start_sequence_id_;
   const uint64_t sequence_id_range_;
   const size_t sequence_length_;
-
- private:
-  std::default_random_engine rng_generator_;
 };
 
 }}  // namespace triton::perfanalyzer
