+++ src/c++/perf_analyzer/test_concurrency_manager.cc
@@ -64,16 +64,17 @@ class TestConcurrencyWorker : public ConcurrencyWorker {
   {
   }
 
-  /// Mock out the InferInput object
-  ///
-  cb::Error CreateInferInput(
-      cb::InferInput** infer_input, const cb::BackendKind kind,
-      const std::string& name, const std::vector<int64_t>& dims,
-      const std::string& datatype) override
-  {
-    *infer_input = new cb::MockInferInput(kind, name, dims, datatype);
-    return cb::Error::Success;
-  }
+  // FIXME going to have to revive this testing
+  //  /// Mock out the InferInput object
+  //  ///
+  //  cb::Error CreateInferInput(
+  //      cb::InferInput** infer_input, const cb::BackendKind kind,
+  //      const std::string& name, const std::vector<int64_t>& dims,
+  //      const std::string& datatype) override
+  //  {
+  //    *infer_input = new cb::MockInferInput(kind, name, dims, datatype);
+  //    return cb::Error::Success;
+  //  }
 };
 
 class MockConcurrencyWorker : public IWorker {
